+++ dec.d.ts
@@ -2,4 +2,5 @@ declare module "react-map-gl-geocoder";
 declare module "as-geo-web-coordinate";
 declare module "lodash.merge";
 declare module "@transak/transak-sdk";
-declare module '@geo-web/mud-world-base-setup'
+declare module "@geo-web/mud-world-base-setup";
+declare module "latlon-geohash";
